src/main/java/com/virtuslab/internship/discount/FifteenPercentDiscount.java
package com.virtuslab.internship.discount;

import com.virtuslab.internship.product.Product;
import com.virtuslab.internship.receipt.Receipt;
import com.virtuslab.internship.receipt.ReceiptEntry;

import java.math.BigDecimal;

public class FifteenPercentDiscount implements Discount{

    private static final String NAME = "FifteenPercentDiscount";
    @Override
    public Receipt apply(Receipt receipt) {
        if (shouldApply(receipt)) {
            var totalPrice = receipt.totalPrice().multiply(BigDecimal.valueOf(0.85));
            var discounts = receipt.discounts();
            discounts.add(NAME);
            return new Receipt(receipt.entries(), discounts, totalPrice);
        }
        return receipt;
    }
    @Override
    public boolean shouldApply(Receipt receipt) {
        int result = receipt.entries()
                .stream()
                .filter(receiptEntry -> receiptEntry.product().type().equals(Product.Type.GRAINS))
                .map(ReceiptEntry::quantity)
                .toList()
                .stream().mapToInt(Integer::intValue).sum();

        return result >= 3;
    }
}
